component AppBar add failing test
